# perf: update IDMS grammar to improve performance