--- a/server/dialect-idms/src/main/antlr4/org/eclipse/lsp/cobol/dialects/idms/IdmsCopyParser.g4
+++ b/server/dialect-idms/src/main/antlr4/org/eclipse/lsp/cobol/dialects/idms/IdmsCopyParser.g4
@@ -15,8 +15,7 @@
 parser grammar IdmsCopyParser;
 options {tokenVocab = IdmsCopyLexer;  superClass = MessageServiceParser;}
 
-startRule: .*? idmsCopybookRules* EOF;
-idmsCopybookRules: (dataDescriptionEntry | copyIdmsStatement | fileDescriptionEntry) .*?;
+startRule: (dataDescriptionEntry | copyIdmsStatement | fileDescriptionEntry | .)* EOF;
 
 dataDescriptionEntry
    : dataDescriptionEntryFormat1