--- a/server/dialect-idms/src/main/antlr4/org/eclipse/lsp/cobol/dialects/idms/IdmsParser.g4
+++ b/server/dialect-idms/src/main/antlr4/org/eclipse/lsp/cobol/dialects/idms/IdmsParser.g4
@@ -15,8 +15,7 @@
 parser grammar IdmsParser;
 options {tokenVocab = IdmsLexer;  superClass = MessageServiceParser;}
 
-startRule: .*? idmsRules* EOF;
-idmsRules: (idmsStatements | idmsSections | idmsIfStatement | ifStatement | copyIdmsStatement) .*?;
+startRule: (idmsStatements | idmsSections | idmsIfStatement | ifStatement | copyIdmsStatement | .)* EOF;
 
 idmsSections
    : idmsControlSection | schemaSection | mapSection