--- a/server/dialect-idms/src/main/antlr4/org/eclipse/lsp/cobol/dialects/idms/IdmsParser.g4
+++ b/server/dialect-idms/src/main/antlr4/org/eclipse/lsp/cobol/dialects/idms/IdmsParser.g4
@@ -15,7 +15,19 @@
 parser grammar IdmsParser;
 options {tokenVocab = IdmsLexer;  superClass = MessageServiceParser;}
 
-startRule: (idmsStatements | idmsSections | idmsIfStatement | ifStatement | copyIdmsStatement | .)* EOF;
+startRule: (
+     (idmsStatements | idmsSections | idmsIfStatement | ifStatement | copyIdmsStatement | ABEND | ACCEPT | ATTACH | BIND
+            | CHANGE | CHECK | COMMIT | CONNECT | COPY | DC | DELETE | DEQUEUE | DISCONNECT | END | ENDPAGE | ENQUEUE
+            | ERASE | FIND | FINISH | FREE | GET | IDMS_CONTROL | IF | INQUIRE | INQUIRE | KEEP | LEVEL_NUMBER | LOAD
+            | MAP | MODIFY | OBTAIN | POST | PUT | READ | READY | RETURN | ROLLBACK | SCHEMA | SEND | SET | SNAP | STARTPAGE
+            | STORE | TRANSFER | WAIT | WRITE
+     )
+     |  ~ (ABEND | ACCEPT | ATTACH | BIND
+         | CHANGE | CHECK | COMMIT | CONNECT | COPY | DC | DELETE | DEQUEUE | DISCONNECT | END | ENDPAGE | ENQUEUE
+         | ERASE | FIND | FINISH | FREE | GET | IDMS_CONTROL | IF | INQUIRE | INQUIRE | KEEP | LEVEL_NUMBER | LOAD
+         | MAP | MODIFY | OBTAIN | POST | PUT | READ | READY | RETURN | ROLLBACK | SCHEMA | SEND | SET | SNAP | STARTPAGE
+         | STORE | TRANSFER | WAIT | WRITE)
+     )* EOF;
 
 idmsSections
    : idmsControlSection | schemaSection | mapSection