--- a/server/dialect-idms/src/main/antlr4/org/eclipse/lsp/cobol/dialects/idms/IdmsTechnicalLexer.g4
+++ b/server/dialect-idms/src/main/antlr4/org/eclipse/lsp/cobol/dialects/idms/IdmsTechnicalLexer.g4
@@ -86,9 +86,6 @@ SQLLINECOMMENT
 	:	SQLLINECOMMENTCHAR ~[\r\n]* NEWLINE -> channel(HIDDEN)
 	;
 
-// treat all the non-processed tokens as errors
-ERRORCHAR : . ;
-
 ZERO_DIGIT: '0';