## Motivation?

Below logs shows IDMS takes a significant long time
```
11:04:04.194 [Analysis thread #0] DEBUG org.eclipse.lsp.cobol.common.benchmark.BenchmarkServiceImpl - ---- Benchmark for uri : zowe-ds:/XXXXXX/XX.XXXX.XXX/XXX
11:04:04.194 [Analysis thread #0] DEBUG org.eclipse.lsp.cobol.common.benchmark.BenchmarkServiceImpl - Timing for Compiler Directives processing: 31,096,900 ns
11:04:04.194 [Analysis thread #0] DEBUG org.eclipse.lsp.cobol.common.benchmark.BenchmarkServiceImpl - Timing for Preprocessing: 271,999,500 ns
11:04:04.194 [Analysis thread #0] DEBUG org.eclipse.lsp.cobol.common.benchmark.BenchmarkServiceImpl - Timing for Dialects processing: 8,092,082,800 ns
11:04:04.195 [Analysis thread #0] DEBUG org.eclipse.lsp.cobol.common.benchmark.BenchmarkServiceImpl - Timing for Transform tree: 53,294,100 ns
11:04:04.195 [Analysis thread #0] DEBUG org.eclipse.lsp.cobol.common.benchmark.BenchmarkServiceImpl - Timing for dialects compiler directives processing: 119,900 ns
11:04:04.195 [Analysis thread #0] DEBUG org.eclipse.lsp.cobol.common.benchmark.BenchmarkServiceImpl - Timing for Implicit dialects processing: 188,500 ns
11:04:04.195 [Analysis thread #0] DEBUG org.eclipse.lsp.cobol.common.benchmark.BenchmarkServiceImpl - Timing for Cleanup document processing: 415,441,600 ns
11:04:04.195 [Analysis thread #0] DEBUG org.eclipse.lsp.cobol.common.benchmark.BenchmarkServiceImpl - Timing for Parsing stage: 48,756,300 ns
``` 

Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration

- [ ] Make sure no regressions are introduced

## Checklist:
- [ ] Each of my commits contains one meaningful change
- [ ] I have performed rebase of my branch on top of the development
- [ ] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] I have checked my code and corrected any misspellings